pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "frameart"
version = "0.1.0"
description = "Generate AI artwork and display it on Samsung Frame TVs"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1,<9",
    "Pillow>=10.0,<11",
    "httpx>=0.27,<1",
    "websocket-client>=1.7,<2",
    "samsungtvws>=2.6,<3",
    "pyyaml>=6.0,<7",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=5.0,<6",
    "ruff>=0.4,<1",
]
api = [
    "fastapi>=0.111,<1",
    "uvicorn[standard]>=0.29,<1",
]

[project.scripts]
frameart = "frameart.cli:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
